refactor: made Deserializer an interface
