+++ src/internal/serialization/src/main/kotlin/kiit/serialization/deserializer/DeserializerPart.kt
@@ -10,13 +10,15 @@ interface DeserializerPart {
 
     /**
      * Given following Source JSON
+     *
      * {
      *       "userId"  : "user123",
      *       "account" : 1234567  ,
      *       "isActive": true
      * }
+     *
      * @param context   : Context info                    : e.g. could be a HttpRequest object
-     * @param parent    : Parent object of current item   : e.g. if "account", this is JSON Root Obj
+     * @param parent    : Parent object of current item   : e.g. if "account", this is JSON Root Obj above
      * @param paramName : Parameter name of  current item : e.g. if "account", this is "account"
      * @param paramValue: Parameter value of current item : e.g. if "account", this is value 1234567
      * @param paramType : Parameter type of current item  : e.g. if "account", this is Int
